+++ lib/screens/checkoutfragments/checkoutProfileWidget.dart
@@ -1,4 +1,3 @@
-
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
@@ -8,14 +7,12 @@ import 'package:ecom/utils/appTheme.dart';
 import 'package:ecom/bloc/bloc_checkout.dart';
 
 class CheckOutProfileScreen extends StatefulWidget {
-
   @override
   _ConfirmOrderPageState createState() => new _ConfirmOrderPageState();
 }
 
-class _ConfirmOrderPageState extends State<CheckOutProfileScreen>  {
-
-  final String REQUIRED="__ *";
+class _ConfirmOrderPageState extends State<CheckOutProfileScreen> {
+  final String REQUIRED = "__ *";
   final _scaffoldKey = GlobalKey<ScaffoldState>();
   @override
   Widget build(BuildContext context) {
@@ -27,16 +24,13 @@ class _ConfirmOrderPageState extends State<CheckOutProfileScreen>  {
   }
 
   Widget _buildBody(BuildContext context) {
-
-    Map map= ModalRoute.of(context).settings.arguments as Map;
-    double amount=map['_amount'];
+    Map map = ModalRoute.of(context).settings.arguments as Map;
+    int amount = map['_amount'];
 
     return Container(
-      padding: EdgeInsets.all(15),
-      child: ProfileInfo(onTap:(){
-        checkOutBloc.selectCheckOut(CheckOutType.SHIPPING);
-      })
-    ) ;
+        padding: EdgeInsets.all(15),
+        child: ProfileInfo(onTap: () {
+          checkOutBloc.selectCheckOut(CheckOutType.SHIPPING);
+        }));
   }
-
 }
